# Make search as fast as original provider