--- a/tests/php/phpstan/phpstan-1.7.8.neon
+++ b/tests/php/phpstan/phpstan-1.7.8.neon
@@ -8,3 +8,4 @@ parameters:
     - '~^Parameter #3 \$groupBy of static method CurrencyCore::getCurrencies\(\) expects bool, Shop given\.$~'
     - '~^Parameter #\d+ \$\w+ of static method ProductCore::priceCalculation\(\) expects \w+, \w+ given\.$~'
     - '~^Access to an undefined property Cookie::\$id_lang\.$~'
+    - '~^Parameter #1 \$master of static method DbCore::getInstance\(\) expects bool, int given\.$~'