| Questions     | Answers
| ------------- | -------------------------------------------------------
| Description?  | Instead of leveraging the search controller, which is not optimal for our usecase, I reimplemented a part of core search functionality to the module itself, which will make it as fast as the original provider.
| Type?         |  improvement
| BC breaks?    | no
| Deprecations? | no
| Fixed ticket? | 
| How to test?  | See below.

**How to test**
- _You don't need to test other pages than search, they were not touched._
- _You don't need to test the filtering in the left column._
- Order of the products could be different compared to when the module is disabled, because the sorting by relevance is not yet implemented.
- **Do test** that the search page finds the same results as if the module is disabled. Try it on 1.7.6 and 8.0.1.

**Example of a performance boost**
Before
- Load Time 3457 ms - You'd better run your shop on a toaster
- Querying Time 2628 ms
- Queries 14950

After
- Load Time 660 ms - Unicorn powered webserver!
- Querying Time 457 ms
- Queries 1505

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/PrestaShop/ps_facetedsearch/789)
<!-- Reviewable:end -->