@sallemiines 2023-02-16
Hello @Hlavtox 

Thanx for the PR , it's LGTM QA :heavy_check_mark: 

**search for a product ( shop without PR )**

![image](https://user-images.githubusercontent.com/31663369/219407242-d50cbd2e-f52b-44e5-8b7d-fe07ae0df8f3.png)

**search for a product ( shop with PR )**

![image](https://user-images.githubusercontent.com/31663369/219407338-cfa24dff-a13c-496b-b33f-c0644e39b42c.png)

Thank you :rocket: